% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mesh.R
\name{plot.intCPUEmesh}
\alias{plot.intCPUEmesh}
\title{Plot an intCPUE mesh}
\usage{
\method{plot}{intCPUEmesh}(x, ...)
}
\arguments{
\item{x}{An intCPUEmesh object}

\item{...}{passed to plot()}
}
\description{
Plot an intCPUE mesh
}
